examples/requirements.txt
kopf
kubernetes
pyyaml
